Another nonce test and reorg
